Fix tests and lint issues
